browse_last {servr}	R Documentation

Reopen the last browsed page

Description

If you have launched a page in the browser via servr but closed it later, you may call this function to reopen it.

Usage

browse_last()

Examples

servr::browse_last()

[Package servr version 0.15 Index]