notebook {servr}	R Documentation

Serve R notebooks

Description

Serve R notebook files created from the rnotebook package.

Usage

notebook(dir = ".", ...)

Arguments

dir
The root directory to serve.

...
server configurations passed to server_config()

Details

An R notebook is basically a JSON file, and the JSON data is inserted to a predefined template in this package (system.file('resources', 'rnotebook.html', package = 'servr')), and rendered through the JavaScript implementation of CommonMark at the moment (the renderer may change in the future).

Examples

# see ?rnotebook::newnb for examples

[Package servr version 0.15 Index]