daemon_stop {servr}	R Documentation

Utilities for daemonized servers

Description

daemon_list() returns IDs of servers, which can be used to stop the daemonized servers.

Usage

daemon_stop(which = daemon_list())

daemon_list()

Arguments

which
A integer vector of the server IDs; by default, IDs of all existing servers in the current R session obtained from daemon_list(), i.e., all daemon servers will be stopped by default.

Value

The function daemon_list() returns a list of existing server IDs, and daemon_stop() returns an invisible NULL.

[Package servr version 0.15 Index]